Move `Question` and `Answer` to separate modules
