+++ TriviaEngine/TriviaEngine/Question.swift
@@ -1,15 +1,3 @@
-import Foundation
-
-public struct Answer: Equatable {
-    public let id: UUID
-    public let text: String
-
-    public init(id: UUID, text: String) {
-        self.id = id
-        self.text = text
-    }
-}
-
 public struct Question: Equatable {
     public let title: String
     public let answers: [Answer]
